requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.10
numpy>=1.20.0
pillow>=8.0.0
pygame>=2.0.0 